pyproject.toml
[tool.poetry]
name = "coding-agent-langgraph"
version = "0.1.0"
description = ""
authors = ["Shuo Yang"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
langchain = "^0.2.7"
langgraph = "^0.1.8"
openai = "^1.35.13"
python-dotenv = "^1.0.1"
langchain-community = "^0.2.7"
langchain-openai = "^0.1.16"
graphviz = "^0.20.3"
networkx = "^3.3"
matplotlib = "^3.9.1"
ipython = "^8.26.0"
pygraphviz = "^1.13"
grandalf = "^0.8"
pytest = "^8.2.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
